Fixed interface data length sizes

+++ include/core/i2c.h
@@ -6,7 +6,7 @@
 uint8_t i2c_init();
 uint8_t i2c_release();
 uint8_t i2c_set_frequency(uint32_t frequency);
-uint8_t i2c_write(uint8_t address, uint8_t reg_address, uint8_t * bytes, uint8_t length);
-uint8_t i2c_read(uint8_t address, uint8_t reg_address, uint8_t * bytes, uint8_t length);
+uint8_t i2c_write(uint8_t address, uint8_t reg_address, uint8_t * bytes, uint32_t length);
+uint8_t i2c_read(uint8_t address, uint8_t reg_address, uint8_t * bytes, uint32_t length);
 
 #endif
